+++ src/DonorProfiles/Admin/UpgradeNotice.php
@@ -29,9 +29,9 @@ class UpgradeNotice {
 			return false;
 		}
 
-		$donorProfilePageIsSet = ! empty( give_get_option( 'donor_profile_page' ) ) && get_post_status( give_get_option( 'donor_profile_page' ) );
-		$historyPageIsSet      = ! empty( give_get_option( 'history_page' ) );
-		return ! $donorProfilePageIsSet && $historyPageIsSet;
+		$donorDashboardPageIsSet = ! empty( give_get_option( 'donor_dashboard_page' ) ) && get_post_status( give_get_option( 'donor_dashboard_page' ) );
+		$historyPageIsSet        = ! empty( give_get_option( 'history_page' ) );
+		return ! $donorDashboardPageIsSet && $historyPageIsSet;
 	}
 
 	/**