+++ src/DonorProfiles/resources/js/app/components/annual-receipt-table/index.js
@@ -45,16 +45,16 @@ const AnnualReceiptTable = ( { annualReceipts, perPage } ) => {
 		<Table
 			header={
 				<Fragment>
-					<div className="give-donor-profile-table__column">
+					<div className="give-donor-dashboard-table__column">
 						{ __( 'Year', 'give' ) }
 					</div>
-					<div className="give-donor-profile-table__column">
+					<div className="give-donor-dashboard-table__column">
 						{ __( 'Amount', 'give' ) }
 					</div>
-					<div className="give-donor-profile-table__column">
+					<div className="give-donor-dashboard-table__column">
 						{ __( 'Count', 'give' ) }
 					</div>
-					<div className="give-donor-profile-table__column">
+					<div className="give-donor-dashboard-table__column">
 						{ __( 'Statement', 'give' ) }
 					</div>
 				</Fragment>
@@ -68,10 +68,10 @@ const AnnualReceiptTable = ( { annualReceipts, perPage } ) => {
 
 			footer={
 				<Fragment>
-					<div className="give-donor-profile-table__footer-text">
+					<div className="give-donor-dashboard-table__footer-text">
 						{ annualReceipts && `${ __( 'Showing', 'give' ) } ${ start + 1 } - ${ end } ${ __( 'of', 'give' ) } ${ annualReceiptsArray.length } ${ __( 'Donations', 'give' ) }` }
 					</div>
-					<div className="give-donor-profile-table__footer-nav">
+					<div className="give-donor-dashboard-table__footer-nav">
 						{ page - 1 >= 1 && (
 							<a onClick={ () => setPage( page - 1 ) }>
 								<FontAwesomeIcon icon="chevron-left" />