+++ src/DonorProfiles/resources/js/app/components/auth-modal/utils/index.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 import { getAPIRoot, getAPINonce } from '../../../utils';
 
 export const loginWithAPI = ( { login, password } ) => {
-	return axios.post( getAPIRoot() + 'give-api/v2/donor-profile/login', {
+	return axios.post( getAPIRoot() + 'give-api/v2/donor-dashboard/login', {
 		login,
 		password,
 	}, 	{
@@ -14,7 +14,7 @@ export const loginWithAPI = ( { login, password } ) => {
 };
 
 export const verifyEmailWithAPI = ( { email, recaptcha } ) => {
-	return axios.post( getAPIRoot() + 'give-api/v2/donor-profile/verify-email', {
+	return axios.post( getAPIRoot() + 'give-api/v2/donor-dashboard/verify-email', {
 		email,
 		'g-recaptcha-response': recaptcha,
 	}, 	{