+++ src/DonorProfiles/resources/js/app/components/divider/index.js
@@ -1,7 +1,7 @@
 import './style.scss';
 const Divider = () => {
 	return (
-		<div className="give-donor-profile-divider" />
+		<div className="give-donor-dashboard-divider" />
 	);
 };
 export default Divider;