+++ src/DonorProfiles/resources/js/app/components/donation-receipt/index.js
@@ -12,18 +12,18 @@ const DonationReceipt = ( { donation } ) => {
 	return receipt.map( ( section, sectionIndex ) => {
 		const lineItems = section.lineItems.map( ( item, itemIndex ) => {
 			return (
-				<div className={ `give-donor-profile-donation-receipt__row${ item.class.includes( 'total' ) ? ' give-donor-profile-donation-receipt__row--footer' : '' }` } key={ itemIndex }>
-					<div className="give-donor-profile-donation-receipt__detail">
+				<div className={ `give-donor-dashboard-donation-receipt__row${ item.class.includes( 'total' ) ? ' give-donor-dashboard-donation-receipt__row--footer' : '' }` } key={ itemIndex }>
+					<div className="give-donor-dashboard-donation-receipt__detail">
 						{ item.icon && <FontAwesomeIcon icon={ item.icon } fixedWidth={ true } /> } { item.label }
 					</div>
-					<div className="give-donor-profile-donation-receipt__value">
+					<div className="give-donor-dashboard-donation-receipt__value">
 						{ item.value }
 					</div>
 				</div>
 			);
 		} );
 		return (
-			<div className="give-donor-profile-donation-receipt__table" key={ sectionIndex }>
+			<div className="give-donor-dashboard-donation-receipt__table" key={ sectionIndex }>
 				{ lineItems }
 			</div>
 		);