+++ src/DonorProfiles/resources/js/app/components/heading/index.js
@@ -4,8 +4,8 @@ import './style.scss';
 
 const Heading = ( { icon, children } ) => {
 	return (
-		<div className="give-donor-profile-heading">
-			{ icon && ( <FontAwesomeIcon icon={ icon } className={ icon === 'spinner' ? 'give-donor-profile-heading__spinner' : '' } /> ) }
+		<div className="give-donor-dashboard-heading">
+			{ icon && ( <FontAwesomeIcon icon={ icon } className={ icon === 'spinner' ? 'give-donor-dashboard-heading__spinner' : '' } /> ) }
 			{ children }
 		</div>
 	);