+++ src/DonorProfiles/resources/js/app/components/logout-modal/utils/index.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 import { getAPIRoot, getAPINonce } from '../../../utils';
 
 export const logoutWithAPI = () => {
-	return axios.post( getAPIRoot() + 'give-api/v2/donor-profile/logout', {
+	return axios.post( getAPIRoot() + 'give-api/v2/donor-dashboard/logout', {
 		headers: {
 			'X-WP-Nonce': getAPINonce(),
 		},