+++ src/DonorProfiles/resources/js/app/components/table/index.js
@@ -2,14 +2,14 @@ import './style.scss';
 
 const Table = ( { header, rows, footer } ) => {
 	return (
-		<div className="give-donor-profile-table">
-			<div className="give-donor-profile-table__header">
+		<div className="give-donor-dashboard-table">
+			<div className="give-donor-dashboard-table__header">
 				{ header }
 			</div>
-			<div className="give-donor-profile-table__rows">
+			<div className="give-donor-dashboard-table__rows">
 				{ rows }
 			</div>
-			<div className="give-donor-profile-table__footer">
+			<div className="give-donor-dashboard-table__footer">
 				{ footer }
 			</div>
 		</div>