+++ src/DonorProfiles/resources/js/app/tabs/donation-history/stats/index.js
@@ -15,48 +15,48 @@ const Stats = () => {
 	const currency = useSelector( ( state ) => state.currency );
 
 	return (
-		<div className="give-donor-profile-dashboard__stats">
+		<div className="give-donor-dashboard-dashboard__stats">
 			{ count && (
-				<div className="give-donor-profile-dashboard__stat">
-					<div className="give-donor-profile-dashboard__figure" style={ { color: accentColor } }>
+				<div className="give-donor-dashboard-dashboard__stat">
+					<div className="give-donor-dashboard-dashboard__figure" style={ { color: accentColor } }>
 						{ count }
 					</div>
-					<div className="give-donor-profile-dashboard__detail">
+					<div className="give-donor-dashboard-dashboard__detail">
 						{ __( 'Number of donations', 'give' ) }
 					</div>
 				</div>
 			) }
 			{ revenue && (
-				<div className="give-donor-profile-dashboard__stat">
-					<div className="give-donor-profile-dashboard__figure" style={ { color: accentColor } }>
+				<div className="give-donor-dashboard-dashboard__stat">
+					<div className="give-donor-dashboard-dashboard__figure" style={ { color: accentColor } }>
 						{ currency.position === 'before' ? (
 							<Fragment>
-								<span className="give-donor-profile-dashboard__figure-currency">{ currency.symbol }</span>{ revenue }
+								<span className="give-donor-dashboard-dashboard__figure-currency">{ currency.symbol }</span>{ revenue }
 							</Fragment>
 						) : (
 							<Fragment>
-								{ revenue }<span className="give-donor-profile-dashboard__figure-currency">{ currency.symbol }</span>
+								{ revenue }<span className="give-donor-dashboard-dashboard__figure-currency">{ currency.symbol }</span>
 							</Fragment>
 						) }
 					</div>
-					<div className="give-donor-profile-dashboard__detail">
+					<div className="give-donor-dashboard-dashboard__detail">
 						{ __( 'Lifetime donations', 'give' ) }
 					</div>
 				</div>
 			) }
 			{ average && (
-				<div className="give-donor-profile-dashboard__stat">
-					<div className="give-donor-profile-dashboard__figure" style={ { color: accentColor } }>
+				<div className="give-donor-dashboard-dashboard__stat">
+					<div className="give-donor-dashboard-dashboard__figure" style={ { color: accentColor } }>
 						{ currency.position === 'before' ? (
 							<Fragment>
-								<span className="give-donor-profile-dashboard__figure-currency">{ currency.symbol }</span>{ average }
+								<span className="give-donor-dashboard-dashboard__figure-currency">{ currency.symbol }</span>{ average }
 							</Fragment>
 						) : (
 							<Fragment>
-								{ average }<span className="give-donor-profile-dashboard__figure-currency">{ currency.symbol }</span>
+								{ average }<span className="give-donor-dashboard-dashboard__figure-currency">{ currency.symbol }</span>
 							</Fragment>
 						) }					</div>
-					<div className="give-donor-profile-dashboard__detail">
+					<div className="give-donor-dashboard-dashboard__detail">
 						{ __( 'Average donation', 'give' ) }
 					</div>
 				</div>